Refactor `slackClient` into single instance
The `SlackClient` instance is shared among all components that interact
with the API. There was little benefit from using dependency injection,
so the `slackClient` module was changed to export a single instance.
This also reduces any possible confusion from considering the client as
a state since object mutations would not necessarily cause a re-render.