Rework login system to fix consistency issues
The login system was reworked to fit a more declarative approach,
removing the need for `useHistory` and `useLocation`.
The previous system had an issue where pages could be forcibly accessed
by manually loading the url.

The new model implements protected routes controlled through states and
redirects the user as needed. The loaded `workspace` is also used to
indicate token validation instead of `isLoading`.

The token input was rebuilt as a controlled component, and an error
message is displayed when the token is invalid. It would be nice to
eventually implement proper error type checking.